refactor: extract mock fixture #109
- users, cards에 해당하는 리소스 mocking을 추출함
