+++ src/api/AxiosClient.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { getCardsAPI, signInAPI } from '.';
 import { STORAGE_KEY } from '@/constant/config';
+import { cards as cardsMock } from '@/mocks/fixture';
 
 describe('', () => {
   afterEach(() => {
@@ -20,24 +21,7 @@ describe('', () => {
     // 리소스를 요청
     const cards = await getCardsAPI();
 
-    expect(cards).toEqual([
-      {
-        _id: '1234asdf',
-        question: '도큐사우르스 짱짱맨',
-        answer: '킹정',
-        submitDate: 'Wed May 17 2023 21:11:26 GMT+0900 (한국 표준시)',
-        stackCount: '0',
-        userId: '1234asdf',
-      },
-      {
-        _id: '1234qwer',
-        question: '블로그를 더 간지나게 만드는 방법',
-        answer: 'github pages로 DIY로 만든다.',
-        submitDate: 'Wed May 17 2023 21:11:26 GMT+0900 (한국 표준시)',
-        stackCount: '0',
-        userId: '1234asdf',
-      },
-    ]);
+    expect(cards).toEqual(cardsMock.documents);
   });
 
   it('should refresh token and try again', async () => {
@@ -53,24 +37,7 @@ describe('', () => {
     const cards = await getCardsAPI();
 
     // 자동 갱신 후 재 요청
-    expect(cards).toEqual([
-      {
-        _id: '1234asdf',
-        question: '도큐사우르스 짱짱맨',
-        answer: '킹정',
-        submitDate: 'Wed May 17 2023 21:11:26 GMT+0900 (한국 표준시)',
-        stackCount: '0',
-        userId: '1234asdf',
-      },
-      {
-        _id: '1234qwer',
-        question: '블로그를 더 간지나게 만드는 방법',
-        answer: 'github pages로 DIY로 만든다.',
-        submitDate: 'Wed May 17 2023 21:11:26 GMT+0900 (한국 표준시)',
-        stackCount: '0',
-        userId: '1234asdf',
-      },
-    ]);
+    expect(cards).toEqual(cardsMock.documents);
     expect(localStorage.getItem(STORAGE_KEY.ACCESS_TOKEN)).not.toBe('expired');
   });
